Spring homework part #2

+++ src/main/java/com/edu/ulab/app/dto/BookDto.java
@@ -8,5 +8,5 @@ public class BookDto {
     private Long userId;
     private String title;
     private String author;
-    private long pageCount;
+    private int pageCount;
 }
